Move MergeDictOperand to mars.operands
