+++ .github/workflows/run-tests.sh
@@ -23,10 +23,10 @@ if [ -z "$NO_COMMON_TESTS" ]; then
     coverage report
   else
     mkdir -p build
-    pytest $PYTEST_CONFIG --cov-config .coveragerc-threaded mars/tensor mars/dataframe mars/web mars/learn
+    pytest $PYTEST_CONFIG --cov-config .coveragerc-threaded mars/tensor mars/dataframe mars/web mars/learn mars/remote
     mv .coverage build/.coverage.tensor.file
     pytest $PYTEST_CONFIG --cov-config .coveragerc --forked --ignore mars/tensor --ignore mars/dataframe \
-      --ignore mars/learn mars
+      --ignore mars/learn --ignore mars/remote mars
     mv .coverage build/.coverage.main.file
     coverage combine build/ && coverage report --fail-under=85
