+++ mars/learn/contrib/lightgbm/train.py
@@ -23,10 +23,10 @@ from functools import reduce
 import numpy as np
 
 from .... import opcodes
+from ....operands import MergeDictOperand, OutputType
 from ....context import get_context, RunningMode
 from ....core import ExecutableTuple
 from ....serialize import DictField, Int32Field, KeyField, ListField, StringField, ValueType
-from ...operands import LearnMergeDictOperand, OutputType
 from ...utils import concat_chunks
 from .align import align_data_set
 from .core import LGBMModelType, get_model_cls_from_type
@@ -34,7 +34,7 @@ from .core import LGBMModelType, get_model_cls_from_type
 logger = logging.getLogger(__name__)
 
 
-class LGBMTrain(LearnMergeDictOperand):
+class LGBMTrain(MergeDictOperand):
     _op_type_ = opcodes.LGBM_TRAIN
 
     _model_type = Int32Field('model_type', on_serialize=lambda x: x.value,
@@ -59,13 +59,13 @@ class LGBMTrain(LearnMergeDictOperand):
     def __init__(self, model_type=None, data=None, label=None, sample_weight=None, init_score=None,
                  eval_datas=None, eval_labels=None, eval_sample_weights=None, eval_init_scores=None,
                  params=None, kwds=None, lgbm_endpoints=None, lgbm_port=None, tree_learner=None,
-                 timeout=None, merge=False, output_types=None, **kw):
+                 timeout=None, **kw):
         super().__init__(_model_type=model_type, _params=params, _data=data, _label=label,
                          _sample_weight=sample_weight, _init_score=init_score, _eval_datas=eval_datas,
                          _eval_labels=eval_labels, _eval_sample_weights=eval_sample_weights,
                          _eval_init_scores=eval_init_scores, _kwds=kwds, _lgbm_endpoints=lgbm_endpoints,
-                         _lgbm_port=lgbm_port, _tree_learner=tree_learner, _timeout=timeout, _merge=merge,
-                         _output_types=output_types, **kw)
+                         _lgbm_port=lgbm_port, _tree_learner=tree_learner, _timeout=timeout,
+                         **kw)
         if self.output_types is None:
             self.output_types = [OutputType.object]
